EISMessageBus: Fix for server-side/publisher-side authentication error
Fix for an authentication error being returned on the server-side of a
request/response and on the publisher-side of publish/subscribe when a client
or subscriber fails to connect their respective endpoints. Prior to this
fix, an error would be returned to the publisher when auth failed for a
subscriber and vice-versa for when a client failed to connect to a
server. These issues are not a problem on the server/publisher side, but on
the client/subscriber side.

Signed-off-by: Kevin Midkiff <kevin.midkiff@intel.com>